addeed confirmation message to delete tweet button

+++ src/components/DeleteTweetButton.js
@@ -1,11 +1,18 @@
 import React, { Component } from 'react'
+import toaster from 'toasted-notes';
 
 class DeleteTweetButton extends Component {
 
     tweeterDeleter= () => {
-        fetch(this.props.apiurl + `/tweets/${this.props.tweet.id}`, {method: 'DELETE'}
-        )
-        this.props.deleteTweetFromState(this.props.tweet)
+        if (window.confirm("Are you sure you want to delete this tweet from the collection?")) {
+            fetch(this.props.apiurl + `/tweets/${this.props.tweet.id}`, {method: 'DELETE'}
+            )
+            this.props.deleteTweetFromState(this.props.tweet)
+            toaster.notify(`Tweet removed from collection!`, {
+                position: 'bottom-left',
+                duration: 2000
+            })
+        }
     }
 
     render() {
